Fix #1: Missing location images on anonymous homepage.
The main issue was to use Pillow for building in
a virtualenv rather than PIL. This branch also
switches out sorl.thumbnail for easy_thumbnails.
sorl added some complexity in setup and debugging
that made it hard to work out what was going on.

Also, images for locations are not saved in tree
for development anymore.