Redesign the homepage for easier development.
This provides a simpler UI, which makes it easier
to build the site. This looks ahead to a simple
event-based site that lists events in an endless
scroller style interface.

Interior pages don't work, and there is no nav
in the header. The goal here is only to get the
start of a simple UI.